+++ config/engine.conf/src/test/java/io/aklivity/zilla/config/engine/test/internal/binding/config/TestBindingOptionsConfigBuilder.java
@@ -117,6 +117,21 @@ public final class TestBindingOptionsConfigBuilder<T> extends ConfigBuilder<T, T
         return this;
     }
 
+    public TestBindingOptionsConfigBuilder<T> authorization(
+        String name,
+        String credentials,
+        String callback,
+        Map<String, String> callbackParams,
+        String expectIdentity,
+        String expectCredentials,
+        Map<String, String> attributes,
+        boolean releaseOnEnd)
+    {
+        this.authorization = new TestAuthorizationConfig(
+            name, credentials, callback, callbackParams, expectIdentity, expectCredentials, attributes, releaseOnEnd);
+        return this;
+    }
+
     public TestBindingOptionsConfigBuilder<T> event(
         long timestamp,
         String message)