+++ runtime/binding-kafka/src/main/java/io/aklivity/zilla/runtime/binding/kafka/internal/stream/KafkaClientAlterConfigsFactory.java
@@ -1159,6 +1159,8 @@ public final class KafkaClientAlterConfigsFactory extends KafkaClientSaslHandsha
 
             doEnd(network, originId, routedId, initialId, initialSeq, initialAck, initialMax,
                 traceId, authorization, EMPTY_EXTENSION);
+
+            deauthorizeGuardSession();
         }
 
         private void doNetworkAbort(
@@ -1173,6 +1175,8 @@ public final class KafkaClientAlterConfigsFactory extends KafkaClientSaslHandsha
 
             cleanupEncodeSlotIfNecessary();
             cleanupBudgetIfNecessary();
+
+            deauthorizeGuardSession();
         }
 
         private void doNetworkReset(
@@ -1186,6 +1190,8 @@ public final class KafkaClientAlterConfigsFactory extends KafkaClientSaslHandsha
             }
 
             cleanupDecodeSlotIfNecessary();
+
+            deauthorizeGuardSession();
         }
 
         private void doNetworkWindow(