+++ runtime/binding-kafka/src/main/java/io/aklivity/zilla/runtime/binding/kafka/internal/stream/KafkaClientCreateTopicsFactory.java
@@ -1231,6 +1231,8 @@ public final class KafkaClientCreateTopicsFactory extends KafkaClientSaslHandsha
 
             doEnd(network, originId, routedId, initialId, initialSeq, initialAck, initialMax,
                 traceId, authorization, EMPTY_EXTENSION);
+
+            deauthorizeGuardSession();
         }
 
         private void doNetworkAbort(
@@ -1245,6 +1247,8 @@ public final class KafkaClientCreateTopicsFactory extends KafkaClientSaslHandsha
 
             cleanupEncodeSlotIfNecessary();
             cleanupBudgetIfNecessary();
+
+            deauthorizeGuardSession();
         }
 
         private void doNetworkReset(
@@ -1258,6 +1262,8 @@ public final class KafkaClientCreateTopicsFactory extends KafkaClientSaslHandsha
             }
 
             cleanupDecodeSlotIfNecessary();
+
+            deauthorizeGuardSession();
         }
 
         private void doNetworkWindow(